+++ lib/pages/ativos_carteira_page.dart
@@ -1,6 +1,6 @@
 import "package:flutter/material.dart";
-import 'package:simulador_investimentos/pages/template/default_statefull_page.dart';
-import 'package:simulador_investimentos/pages/template/default_statefull_page_state.dart';
+import 'package:simulador_investimentos/pages/base/default_statefull_page.dart';
+import 'package:simulador_investimentos/pages/base/default_statefull_page_state.dart';
 import 'package:simulador_investimentos/widgets/card_ativos_carteira.dart';
 
 class AtivosCarteiraPage extends DefaultStatefullPage {
