+++ lib/pages/base/base_loadable_bloc.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:rxdart/rxdart.dart';
-import 'package:simulador_investimentos/pages/bloc_events.dart';
+import 'package:simulador_investimentos/pages/base/bloc_events.dart';
 
 import 'base_bloc.dart';
