+++ lib/pages/operacao_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:simulador_investimentos/core/model/domain/ativo.dart';
 import 'package:simulador_investimentos/core/model/domain/tipo_operacao.dart';
-import 'package:simulador_investimentos/pages/template/default_statefull_page.dart';
-import 'package:simulador_investimentos/pages/template/default_statefull_page_state.dart';
+import 'package:simulador_investimentos/pages/base/default_statefull_page.dart';
+import 'package:simulador_investimentos/pages/base/default_statefull_page_state.dart';
 import 'package:simulador_investimentos/widgets/card_operacao.dart';
 
 class OperacaoPage extends DefaultStatefullPage {
